+++ src/main/java/projet/studenity/service/ProductService.java
@@ -7,10 +7,10 @@ import java.util.List;
 public interface ProductService {
     public abstract boolean createProduct(Product product);
     public abstract boolean updateProduct(Product product);
-    public abstract boolean deleteProduct(Long id);
+    public abstract boolean deleteProduct(int id);
     public abstract List<Product> getProducts();
-    public Product findProductById(long id);
+    public Product findProductById(int id);
     public List<Product> findProductByName(String firstName);
-    public boolean reserveProduct(Long id);
-    public List<Product> findProductByCategory(Long idCategory);
+    public boolean reserveProduct(int id);
+    public List<Product> findProductByCategory(int idCategory);
 }
